+++ src/main/java/com/safetypin/post/service/PostService.java
@@ -1,35 +1,232 @@
 package com.safetypin.post.service;
 
-import com.safetypin.post.dto.FeedQueryDTO;
-import com.safetypin.post.dto.PostCreateRequest;
-import com.safetypin.post.dto.PostedByData;
+import com.safetypin.post.dto.*;
+import com.safetypin.post.exception.InvalidPostDataException;
+import com.safetypin.post.exception.PostException;
+import com.safetypin.post.exception.PostNotFoundException;
+import com.safetypin.post.exception.UnauthorizedAccessException;
+import com.safetypin.post.model.Category;
 import com.safetypin.post.model.Post;
+import com.safetypin.post.repository.CategoryRepository;
+import com.safetypin.post.repository.PostRepository;
+import com.safetypin.post.service.strategy.DistanceFeedStrategy;
+import com.safetypin.post.service.strategy.FeedStrategy;
+import com.safetypin.post.service.strategy.TimestampFeedStrategy;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
+@Slf4j
 @Service
-public interface PostService {
+public class PostService {
 
-    Post createPost(PostCreateRequest postCreateRequest);
+    private final PostRepository postRepository;
+    private final CategoryRepository categoryRepository;
+    private final DistanceFeedStrategy distanceFeedStrategy;
+    private final TimestampFeedStrategy timestampFeedStrategy;
+    @Value("${be-auth}")
+    private String apiEndpoint = "http://safetypin.ppl.cs.ui.ac.id";
 
-    List<Post> findAll(); // debugging purposes
+    @Autowired
+    public PostService(PostRepository postRepository, CategoryRepository categoryRepository,
+                       DistanceFeedStrategy distanceFeedStrategy,
+                       TimestampFeedStrategy timestampFeedStrategy) {
+        this.postRepository = postRepository;
+        this.categoryRepository = categoryRepository;
+        this.distanceFeedStrategy = distanceFeedStrategy;
+        this.timestampFeedStrategy = timestampFeedStrategy;
+    }
 
-    Page<Post> findAllPaginated(Pageable pageable);
+    private static List<LinkedHashMap<String, String>> getLinkedHashMaps(ResponseEntity<AuthResponse> result) {
+        Object data = Objects.requireNonNull(result.getBody()).getData();
+        List<?> rawList = (List<?>) data;
+        List<LinkedHashMap<String, String>> linkedHashMapList = new ArrayList<>();
 
-    Post findById(UUID id);
+        for (Object obj : rawList) {
+            if (obj instanceof LinkedHashMap<?, ?> map) {
+                LinkedHashMap<String, String> stringMap = new LinkedHashMap<>();
+                for (Map.Entry<?, ?> entry : map.entrySet()) {
+                    stringMap.put(
+                            String.valueOf(entry.getKey()),
+                            String.valueOf(entry.getValue())  // force String conversion
+                    );
+                }
+                linkedHashMapList.add(stringMap);
+            }
+        }
+        return linkedHashMapList;
+    }
 
-    Page<Map<String, Object>> findPostsByUser(UUID userId, Pageable pageable);
+    // find all (debugging purposes)
 
-    // New method to delete a post by ID, with user validation
-    void deletePost(UUID postId, UUID userId);
+    public List<Post> findAll() {
+        return postRepository.findAll();
+    }
 
-    // Method using strategy pattern for feed algorithms
-    Page<Map<String, Object>> getFeed(FeedQueryDTO queryDTO, String feedType);
+    // find all with pagination
 
-    List<PostedByData> fetchProfiles();
+    public Page<Post> findAllPaginated(Pageable pageable) {
+        return postRepository.findAll(pageable);
+    }
+
+
+    public Post createPost(PostCreateRequest postCreateRequest) {
+
+        String title = postCreateRequest.getTitle();
+        String content = postCreateRequest.getCaption();
+        Double latitude = postCreateRequest.getLatitude();
+        Double longitude = postCreateRequest.getLongitude();
+        String category = postCreateRequest.getCategory();
+        UUID postedBy = postCreateRequest.getPostedBy();
+        String imageUrl = postCreateRequest.getImageUrl();
+        String address = postCreateRequest.getAddress();
+
+        if (title == null || title.trim().isEmpty()) {
+            throw new InvalidPostDataException("Title is required");
+        }
+        if (content == null || content.trim().isEmpty()) {
+            throw new InvalidPostDataException("Content is required");
+        }
+        if (latitude == null || longitude == null) {
+            throw new InvalidPostDataException("Location coordinates are required");
+        }
+        if (category == null || category.trim().isEmpty()) {
+            throw new InvalidPostDataException("Category is required");
+        }
+        if (postedBy == null) {
+            throw new InvalidPostDataException("User ID (postedBy) is required");
+        }
+
+        // Verify that the category exists
+        Category categoryObj = categoryRepository.findByName(category);
+        if (categoryObj == null) {
+            throw new InvalidPostDataException("Category does not exist: " + category);
+        }
+
+        // Create the post
+        Post post = new Post.Builder()
+                .title(title)
+                .caption(content)
+                .location(latitude, longitude)
+                .category(category)
+                .postedBy(postedBy) // Set the postedBy value
+                .imageUrl(imageUrl)
+                .address(address)
+                .build();
+
+        try {
+            return postRepository.save(post);
+        } catch (Exception e) {
+            log.error("Error saving post: {}", e.getMessage());
+            throw new PostException("Failed to save the post: " + e.getMessage());
+        }
+    }
+
+
+    public Post findById(UUID id) {
+        return postRepository.findById(id)
+                .orElseThrow(() -> new PostNotFoundException("Post not found with id: " + id));
+    }
+
+
+    public void deletePost(UUID postId, UUID userId) {
+        Post post = findById(postId);
+        if (!post.getPostedBy().equals(userId)) {
+            throw new UnauthorizedAccessException("User not authorized to delete this post");
+        }
+        postRepository.delete(post);
+    }
+
+
+    public Page<Map<String, Object>> getFeed(FeedQueryDTO queryDTO, String feedType) {
+        // Validate categories if provided
+        if (queryDTO.getCategories() != null && !queryDTO.getCategories().isEmpty()) {
+            validateCategories(queryDTO.getCategories());
+        }
+
+        // Validate feed type
+        if (feedType == null) {
+            throw new IllegalArgumentException("Feed type is required");
+        }
+
+        // Choose strategy based on feed type
+        FeedStrategy strategy = switch (feedType.toLowerCase()) {
+            case "distance" -> distanceFeedStrategy;
+            case "timestamp" -> timestampFeedStrategy;
+            default -> throw new IllegalArgumentException("Invalid feed type: " + feedType);
+        };
+
+        // Get all posts
+        List<Post> allPosts = postRepository.findAll();
+
+        // collect createdBy UUID from allPosts
+        List<UUID> createdByList = allPosts.stream()
+                .map(Post::getPostedBy)
+                .distinct()
+                .toList();
+
+        // fetch profiles
+        Map<UUID, PostedByData> profileList = fetchPostedByData(createdByList);
+
+        // Apply strategy to posts
+        return strategy.processFeed(allPosts, queryDTO, profileList);
+    }
+
+
+    public Page<Map<String, Object>> findPostsByUser(UUID postUserId, Pageable pageable) {
+        if (postUserId == null) {
+            throw new IllegalArgumentException("Post user ID is required");
+        }
+
+        // Get all posts with filters and ordered
+        Page<Post> allPosts = postRepository.findByPostedByOrderByCreatedAtDesc(postUserId, pageable);
+
+        // collect createdBy UUID from allPosts
+        List<UUID> createdByList = allPosts.stream()
+                .map(Post::getPostedBy)
+                .distinct()
+                .toList();
+        // fetch profiles
+        Map<UUID, PostedByData> profileList = fetchPostedByData(createdByList);
+
+        // Map to PostData and return page
+        return allPosts.map(post -> {
+            Map<String, Object> result = new HashMap<>();
+            PostData postData = PostData.fromPostAndUserId(post, postUserId, profileList.get(post.getPostedBy()));
+            result.put("post", postData);
+            return result;
+        });
+    }
+
+    // Helper method to validate that all categories exist
+    private void validateCategories(List<String> categories) {
+        for (String category : categories) {
+            Category categoryObj = categoryRepository.findByName(category);
+            if (categoryObj == null) {
+                throw new InvalidPostDataException("Category does not exist: " + category);
+            }
+        }
+    }
+
+    public Map<UUID, PostedByData> fetchPostedByData(List<UUID> userIds) {
+        // fetch profiles
+        RestTemplate restTemplate = new RestTemplate();
+        String uri = apiEndpoint + "/api/profiles/batch"; // or any other uri
+        HttpEntity<List<UUID>> entity = new HttpEntity<>(userIds, null);
+        ResponseEntity<Map<UUID, PostedByData>> result = restTemplate.exchange(uri, HttpMethod.POST, entity, new ParameterizedTypeReference<Map<UUID, PostedByData>>() {
+        });
+
+        log.info("hasil fetch: " + result.getBody() + "  " + result.getStatusCode());
+        return result.getBody();
+    }
 }
